pyproject.toml
[tool.poetry]
name = "arlo-oss-api-client"
version = "ccf05e5f-96ee-4ded-9f2f-6a165f5352c4"
description = "A client library for accessing Arlo OSS API"

authors = []

readme = "README.md"
packages = [
    {include = "arlo_oss_api_client"},
]
include = ["CHANGELOG.md", "arlo_oss_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.6"
httpx = "^0.15.0"
attrs = "^20.1.0"
python-dateutil = "^2.8.1"

[tool.black]
line-length = 120
target_version = ['py36', 'py37', 'py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"